Actualizacion de pagos parciales modulo Ventas

+++ app/Http/Livewire/Admin/Empresas/ShowEmpresa.php
@@ -5,11 +5,11 @@ namespace App\Http\Livewire\Admin\Empresas;
 use App\Helpers\FormatoPersonalizado;
 use App\Helpers\GetClient;
 use App\Models\Empresa;
-use App\Models\Image as Logo;
 use App\Models\Telephone;
 use App\Models\Ubigeo;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 use Livewire\Component;
 use Livewire\WithFileUploads;
 use Intervention\Image\ImageManagerStatic as Image;
@@ -21,7 +21,7 @@ class ShowEmpresa extends Component
 
     public $openphone = false;
     public $empresa, $telephone, $phone, $telefono;
-    public $icono, $logo, $idlogo, $idicono;
+    public $icono, $logo, $idlogo, $idicono, $usemarkagua, $mark;
     public $validatemail;
 
     protected function rules()
@@ -38,7 +38,28 @@ class ShowEmpresa extends Component
             'empresa.igv' => ['required', 'numeric', 'decimal:0,4', 'min:0'],
             'logo' => ['nullable', 'file', 'mimes:jpg,bmp,png'],
             'icono' => ['nullable', 'file', 'mimes:ico'],
+            'mark' => [
+                'nullable',
+                Rule::requiredIf($this->empresa->usarMarkagua() && empty($this->empresa->markagua)),
+                'file', 'mimes:png'
+            ],
             'empresa.uselistprice' => ['integer', 'min:0', 'max:1'],
+            'empresa.viewpriceantes' => ['integer', 'min:0', 'max:1'],
+            'empresa.viewlogomarca' => ['integer', 'min:0', 'max:1'],
+            'empresa.viewtextopromocion' => ['integer', 'min:0', 'max:2'],
+            'empresa.usemarkagua' => ['integer', 'min:0', 'max:1'],
+            'empresa.alignmark' => [
+                'nullable', Rule::requiredIf($this->empresa->usarMarkagua()),
+                'string', 'max:25'
+            ],
+            'empresa.widthmark' => [
+                'nullable', Rule::requiredIf($this->empresa->usarMarkagua()),
+                'integer', 'min:50', 'max:300'
+            ],
+            'empresa.heightmark' => [
+                'nullable', Rule::requiredIf($this->empresa->usarMarkagua()),
+                'integer', 'min:50', 'max:300'
+            ],
             'empresa.usepricedolar' => ['integer', 'min:0', 'max:1'],
             'empresa.tipocambio' => ['nullable', 'required_if:usepricedolar,1', 'numeric', 'decimal:0,4', 'min:0', 'gt:0'],
             'empresa.viewpricedolar' => ['integer', 'min:0', 'max:1', 'numeric', 'decimal:0,4', 'min:0'],
@@ -63,26 +84,23 @@ class ShowEmpresa extends Component
 
     public function update()
     {
-        $this->empresa->uselistprice = $this->empresa->uselistprice == 1 ? 1 : 0;
-        $this->empresa->usepricedolar = $this->empresa->usepricedolar == true ?  1 : 0;
-        $this->empresa->viewpricedolar = $this->empresa->viewpricedolar == true ?  1 : 0;
-        $this->empresa->tipocambioauto = $this->empresa->tipocambioauto == true ?  1 : 0;
-
-        if ($this->empresa->usepricedolar == 0) {
-            $this->empresa->usepricedolar = 0;
-            $this->empresa->viewpricedolar = 0;
-            $this->empresa->tipocambioauto = 0;
-            $this->empresa->tipocambio = null;
-        }
 
         $this->empresa->document = trim($this->empresa->document);
         $this->empresa->name = trim($this->empresa->name);
         $this->empresa->direccion = trim($this->empresa->direccion);
         $this->empresa->estado = trim($this->empresa->estado);
         $this->empresa->condicion = trim($this->empresa->condicion);
-        $this->empresa->montoadelanto = empty($this->empresa->montoadelanto) ? null : $this->empresa->montoadelanto;
-        $this->validate();
-
+        $this->validate([
+            'empresa.document' => ['required', 'numeric', 'digits:11', 'regex:/^\d{11}$/'],
+            'empresa.name' => ['required', 'string', 'min:3'],
+            'empresa.direccion' => ['required', 'string', 'min:3'],
+            'empresa.ubigeo_id' => ['required', 'integer', 'min:0', 'exists:ubigeos,id'],
+            'empresa.estado' => ['required', 'string'],
+            'empresa.condicion' => ['required', 'string'],
+            'empresa.email' => ['nullable', 'email'],
+            'empresa.web' => ['nullable', 'starts_with:http://,https://,https://www.,http://www.,www.'],
+            'empresa.igv' => ['required', 'numeric', 'decimal:0,4', 'min:0'],
+        ]);
         try {
             DB::beginTransaction();
             // if ($this->publickey) {
@@ -94,9 +112,101 @@ class ShowEmpresa extends Component
 
             $this->empresa->email = $this->empresa->email;
             $this->empresa->web = $this->empresa->web;
-            $this->empresa->montoadelanto = $this->empresa->montoadelanto;
             $this->empresa->default = 1;
             $this->empresa->ubigeo_id = $this->empresa->ubigeo_id;
+            $this->empresa->save();
+            DB::commit();
+            $this->resetValidation();
+            $this->dispatchBrowserEvent('updated');
+            $this->empresa->refresh();
+        } catch (\Exception $e) {
+            DB::rollBack();
+            throw $e;
+        } catch (\Throwable $e) {
+            DB::rollBack();
+            throw $e;
+        }
+    }
+
+    public function updateopciones()
+    {
+        $this->empresa->montoadelanto = empty($this->empresa->montoadelanto) ? null : $this->empresa->montoadelanto;
+        $this->empresa->uselistprice = $this->empresa->uselistprice == 1 ? 1 : 0;
+        $this->empresa->usepricedolar = $this->empresa->usepricedolar == true ?  1 : 0;
+        $this->empresa->viewpricedolar = $this->empresa->viewpricedolar == true ?  1 : 0;
+        $this->empresa->tipocambioauto = $this->empresa->tipocambioauto == true ?  1 : 0;
+        $this->empresa->viewpriceantes = $this->empresa->viewpriceantes == true ?  1 : 0;
+        $this->empresa->viewlogomarca = $this->empresa->viewlogomarca == true ?  1 : 0;
+        $this->empresa->usemarkagua = $this->empresa->usemarkagua == true ?  1 : 0;
+
+        if ($this->empresa->usepricedolar == 0) {
+            $this->empresa->usepricedolar = 0;
+            $this->empresa->viewpricedolar = 0;
+            $this->empresa->tipocambioauto = 0;
+            $this->empresa->tipocambio = null;
+        }
+
+        $this->validate([
+            'logo' => ['nullable', 'file', 'mimes:jpg,bmp,png'],
+            'icono' => ['nullable', 'file', 'mimes:ico'],
+            'mark' => [
+                'nullable',
+                Rule::requiredIf($this->empresa->usarMarkagua() && empty($this->empresa->markagua)),
+                'file', 'mimes:png'
+            ],
+            'empresa.uselistprice' => ['integer', 'min:0', 'max:1'],
+            'empresa.viewpriceantes' => ['integer', 'min:0', 'max:1'],
+            'empresa.viewlogomarca' => ['integer', 'min:0', 'max:1'],
+            'empresa.viewtextopromocion' => ['integer', 'min:0', 'max:2'],
+            'empresa.usemarkagua' => ['integer', 'min:0', 'max:1'],
+            'empresa.alignmark' => [
+                'nullable', Rule::requiredIf($this->empresa->usarMarkagua()),
+                'string', 'max:25'
+            ],
+            'empresa.widthmark' => [
+                'nullable', Rule::requiredIf($this->empresa->usarMarkagua()),
+                'integer', 'min:50', 'max:300'
+            ],
+            'empresa.heightmark' => [
+                'nullable', Rule::requiredIf($this->empresa->usarMarkagua()),
+                'integer', 'min:50', 'max:300'
+            ],
+            'empresa.usepricedolar' => ['integer', 'min:0', 'max:1'],
+            'empresa.tipocambio' => ['nullable', 'required_if:usepricedolar,1', 'numeric', 'decimal:0,4', 'min:0', 'gt:0'],
+            'empresa.viewpricedolar' => ['integer', 'min:0', 'max:1', 'numeric', 'decimal:0,4', 'min:0'],
+            'empresa.tipocambioauto' => ['integer', 'min:0', 'max:1'],
+            'empresa.montoadelanto' => ['nullable', 'numeric', 'min:0', 'decimal:0,2'],
+
+        ]);
+
+        try {
+            DB::beginTransaction();
+
+            if ($this->mark) {
+                if (!Storage::directoryExists('images/company/')) {
+                    Storage::makeDirectory('images/company/');
+                }
+
+                $compressedImage = Image::make($this->mark->getRealPath())
+                    ->orientate()->encode('jpg', 70);
+
+                $markURL = uniqid('markagua_') . '.' . $this->mark->getClientOriginalExtension();
+                $compressedImage->save(public_path('storage/images/company/' . $markURL));
+
+                if ($compressedImage->filesize() > 1048576) { //1MB
+                    $compressedImage->destroy();
+                    $compressedImage->delete();
+                    $this->addError('mark', 'La imagen excede el tamaño máximo permitido.');
+                    return false;
+                }
+
+                if ($this->empresa->markagua) {
+                    Storage::delete($this->empresa->getMarkAguaURL());
+                }
+
+                $this->empresa->markagua = $markURL;
+            }
+
             $this->empresa->save();
             DB::commit();
             $this->resetValidation();
@@ -154,30 +264,30 @@ class ShowEmpresa extends Component
         $this->dispatchBrowserEvent('deleted');
     }
 
-    public function deleteicono(Empresa $empresa)
+    public function deleteicono()
     {
-        if ($empresa) {
-            if (Storage::exists('images/company/' . $empresa->icono)) {
-                Storage::delete('images/company/' . $empresa->icono);
+        if ($this->empresa->icono) {
+            if (Storage::exists('images/company/' . $this->empresa->icono)) {
+                Storage::delete('images/company/' . $this->empresa->icono);
+                $this->empresa->icono = null;
+                $this->empresa->save();
+                $this->empresa->refresh();
+                $this->idicono = rand();
+                $this->dispatchBrowserEvent('deleted');
             }
-
-            $empresa->icono = null;
-            $empresa->save();
-            $this->empresa->refresh();
-            $this->idicono = rand();
         }
     }
 
-    public function deletelogo(Logo $logo)
+    public function deletelogo()
     {
-        if ($logo) {
-            if (Storage::exists('images/company/' . $logo->url)) {
-                Storage::delete('images/company/' . $logo->url);
+        if ($this->empresa->image) {
+            if (Storage::exists('images/company/' . $this->empresa->image->url)) {
+                Storage::delete('images/company/' . $this->empresa->image->url);
+                $this->empresa->image->delete();
+                $this->empresa->refresh();
+                $this->idlogo = rand();
+                $this->dispatchBrowserEvent('deleted');
             }
-
-            $logo->deleteOrFail();
-            $this->empresa->refresh();
-            $this->idlogo = rand();
         }
     }
 
@@ -225,24 +335,25 @@ class ShowEmpresa extends Component
         }
 
         if ($this->logo) {
-            if ($this->empresa->image) {
-                $this->empresa->image->deleteOrFail();
-                Storage::delete('images/company/' . $this->empresa->image->url);
-            }
-
             $compressedImage = Image::make($this->logo->getRealPath())
-                ->resize(300, 300, function ($constraint) {
-                    $constraint->aspectRatio();
-                    $constraint->upsize();
-                })->orientate()->encode('jpg', 30);
+                ->orientate()->encode('jpg', 70);
 
-            $urlLogo = uniqid() . '.' . $this->logo->getClientOriginalExtension();
+            $urlLogo = 'logo_' . uniqid() . '.' . $this->logo->getClientOriginalExtension();
             $compressedImage->save(public_path('storage/images/company/' . $urlLogo));
 
             if ($compressedImage->filesize() > 1048576) { //1MB
                 $compressedImage->destroy();
                 $compressedImage->delete();
                 $this->addError('logo', 'La imagen excede el tamaño máximo permitido.');
+                return false;
+            }
+
+
+            if ($this->empresa->image) {
+                if (Storage::exists('images/company/' . $this->empresa->image->url)) {
+                    Storage::delete('images/company/' . $this->empresa->image->url);
+                    $this->empresa->image->delete();
+                }
             }
 
             $this->empresa->image()->create([
@@ -268,11 +379,19 @@ class ShowEmpresa extends Component
         $urlicono = $this->empresa->icono ?? null;
         if ($this->icono) {
             $extpublic = FormatoPersonalizado::getExtencionFile($this->icono->getClientOriginalName());
-            $urlicono = uniqid() . '.' . $extpublic;
+            $urlicono = 'icon_' . uniqid() . '.' . $extpublic;
             Storage::putFileAs('images/company', $this->icono, $urlicono);
             if ($this->icono->getSize() > 1048576) { //1MB
                 $this->addError('icono', 'La imagen excede el tamaño máximo permitido.');
+                return false;
             }
+
+            if ($this->empresa->icono) {
+                if (Storage::exists('images/company/' . $this->empresa->icono)) {
+                    Storage::delete('images/company/' . $this->empresa->icono);
+                }
+            }
+
             $this->empresa->icono = $urlicono;
             $this->empresa->save();
             $this->reset(['icono', 'idicono']);
@@ -303,5 +422,21 @@ class ShowEmpresa extends Component
             $this->addError('logo', $e->getMessage());
             return;
         }
+    }
+
+    public function updatedMark($file)
+    {
+        try {
+            $url = $file->temporaryUrl();
+        } catch (\Exception $e) {
+            $this->reset(['mark']);
+            $this->addError('mark', $e->getMessage());
+            return;
+        }
+    }
+
+    public function clearMark()
+    {
+        $this->reset(['mark']);
     }
 }
